Fix Green LED Not Working BUG
Signed-off-by: Shoaib0597 <Shoaib0595@gmail.com>

Note---
After a thorough review of LED's DT-Entries and Codes, I became sure that these were alright. Since the Green LED seemed to be using PWM for Working, I decided to replace the PWM Version of this Kernel with the one available in Stock Kernel's (Marshmallow) Source. This Change, coupled with Un-Touched Import of LED-related Codes, made the Green LED to work properly.

For reference---
https://github.com/YUTeleventures/kernel_yu_yurekablack